Optimize `Stream_write` and `UDPSocket_send`
From libuv:

 > The buffers to be written must remain valid until the callback is called.
 > This is not required for the uv_buf_t array.

So uv_buf_t has been removed from `WriteReq` and `SendReq`, since the
write and send functions will copy the buffer(s) internally (in a field
named bufsml).
The memory pointed to by the buffers must still survive the function
though, so it is still allocated on the heap alongside the request.